Revert "review changes: reorganize transport credentials and non-temporary error"
This reverts commit 2c90d35cd6626440fa88a95908940552d546cfd8.
