+++ pkg/peertls/tlsopts/tls.go
@@ -4,8 +4,11 @@
 package tlsopts
 
 import (
+	"context"
 	"crypto/tls"
 	"crypto/x509"
+	"net"
+	"strings"
 
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
@@ -13,8 +16,13 @@ import (
 	"storj.io/storj/pkg/identity"
 	"storj.io/storj/pkg/peertls"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/pkg/transport"
 )
+
+// TransportCredentials wraps an embedded credentials.TransportCredentials.
+// It decorates the `ClientHandshake` method.
+type TransportCredentials struct {
+	credentials.TransportCredentials
+}
 
 // ServerOption returns a grpc `ServerOption` for incoming connections
 // to the node with this full identity.
@@ -44,10 +52,10 @@ func (opts *Options) DialOption(id storj.NodeID) grpc.DialOption {
 	return grpc.WithTransportCredentials(opts.TransportCredentials(id))
 }
 
-// Credentials returns a grpc `credentials.Credentials`
+// TransportCredentials returns a grpc `credentials.TransportCredentials`
 // implementation for use within peertls.
 func (opts *Options) TransportCredentials(id storj.NodeID) credentials.TransportCredentials {
-	return &transport.Credentials{
+	return &TransportCredentials{
 		TransportCredentials: credentials.NewTLS(opts.TLSConfig(id)),
 	}
 }
@@ -68,6 +76,19 @@ func (opts *Options) TLSConfig(id storj.NodeID) *tls.Config {
 			pcvFuncs...,
 		),
 	}
+}
+
+// ClientHandshake returns a non-temporary error if an error is returned from
+// the underlying `ClientHandshake` call.
+func (tc *TransportCredentials) ClientHandshake(ctx context.Context, authority string, conn net.Conn) (net.Conn, credentials.AuthInfo, error) {
+	tlsConn, authInfo, err := tc.TransportCredentials.ClientHandshake(ctx, authority, conn)
+	if err != nil {
+		isCertError := peertls.ErrVerifyPeerCert.Has(err) || strings.Contains(err.Error(), "bad certificate")
+		if isCertError {
+			return tlsConn, authInfo, peertls.NewNonTemporaryError(err)
+		}
+	}
+	return tlsConn, authInfo, err
 }
 
 func verifyIdentity(id storj.NodeID) peertls.PeerCertVerificationFunc {
